+++ src/gfx/graphics/binding_set.hpp
@@ -57,7 +57,7 @@ template<typename T>
 void binding_set::bind(u32 binding, const host_buffer<T>& buffer, u32 count, u32 first)
 {
 	assert(_layout->types()[binding] == binding_type::uniform_buffer || _layout->types()[binding] == binding_type::storage_buffer);
-	implementation()->bind(binding, 0, _layout->types()[binding], &buffer.implementation(), first * sizeof(T), count ==  ~0 ? buffer.capacity() * sizeof(T) : count * sizeof(T));
+	implementation()->bind(binding, 0, _layout->types()[binding], &buffer.implementation(), static_cast<u32>(first * sizeof(T)), static_cast<u32>(count ==  ~0 ? buffer.capacity() * sizeof(T) : count * sizeof(T)));
 }
 
 template<typename T>
